refactor: #1497 Rename from Allophone to Sound
- Refactor package name for controllers
